refactor: create thunderstore package
