+++ src-vue/src/utils/thunderstore/ThunderstoreMod.d.ts
